XMLHttpRequest/response-error.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>XMLHttpRequest Test: reponse - error flag set</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="assert" content="Check if XMLHttpRequest.reponse return null if error flag is set">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

async_test(function (t) {
  var client = new XMLHttpRequest();
  client.responseType = "blob";

  client.onerror = t.step_func(function () {
    assert_equals(client.response, null);
    t.done();
  });

  client.open("GET", "resources/not_exist_file.html")
  client.send("null");
}, document.title);

</script>
